VIRO-57 [skeletal]: bake pivot transforms
FBX has a number of transforms that Viro (and
most real-time engines) do not support, like
rotation offset, scale offset, pivot, etc.

Fortunately there is a method ResetPivotSetAndConvertAnimation
that bakes all these transforms into the standard
TRS (translation rotation scale) matrices. Unfortunately,
this method doesn’t work correctly for pre-rotation in FBX
2018. However, it does work in FBX 2015, so this commit
downgrades to that API and implements the method.

It appears to be a bug on Autodesk’s 2018 API that this
method doesn’t work as it’s supposed to.